requirements.txt
bcrypt
base64
PyQt5
configparser
secrets
pycryptodome